+++ docs/release_notes.rst
@@ -1,6 +1,12 @@
 Release Notes
 =============
 
+v2.2.0
+------
+* More optimizations for unseen events
+* [!!!BREAKING!!!] Renamed `get_filtered_events_queries` to `get_filtered_events_queryset` on the `Medium` model and it now returns an `Event` queryset
+* [!!!BREAKING!!!] Querying for events through a medium now only returns events that have a source that are subscribed to that medium.
+
 v2.1.0
 ------
 * Optimize querying for unseen events